Extend translation manager instead of patching it
