+++ home/management/commands/load_po_files.py
@@ -1,16 +1,11 @@
 from django.core.management.base import BaseCommand
 from django.utils.translation import gettext_lazy as _
 
-from translation_manager.manager import Manager
-
-from iogt.patch import patch_store_to_db
+from home.translation_manager.manager import IogtTranslationManager as Manager
 
 
 class Command(BaseCommand):
     def handle(self, *args, **options):
         self.stdout.write(_('Loading translations...'))
-        patch_store_to_db()
-
-        manager = Manager()
-        manager.load_data_from_po()
+        Manager().load_data_from_po()
         self.stdout.write(_('The translations have been loaded successfully!'))
